\name{WeightsNUTS2}
\alias{WeightsNUTS2}

\title{Spatial Weights for European Union NUTS2 Regions}

\description{Spatial weight matrices for NUTS2 regions in the European Union.}

\usage{data("WeightsNUTS2")}

\format{
A list containing 40 listw weight matrices.
}

\source{
  Crespo Cuaresma J, Feldkircher M (2013).
    Spatial Filtering, Model Uncertainty and the Speed of Income Convergence in Europe (Replication Data).
    Version: 1. \emph{Journal of Applied Econometrics}. Dataset.
    \doi{10.15456/jae.2022320.0732268890}
}

\references{
  Crespo Cuaresma J, Feldkircher M (2013).
    Spatial Filtering, Model Uncertainty and the Speed of Income Convergence in Europe.
    \emph{Journal of Applied Econometrics}, \bold{28}(4), 720-741.
    \doi{10.1002/jae.2277}
}

\keyword{datasets}
